requirements.txt
coverage==5.1
Django==2.2.10
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
xlrd==1.2.0
